Fixed a bug in precompiler that caused segmentation fault in 64 bits. (cherry picked from commit dd29cbfbff6e1d8d00fb4e345d2ddb976932ad95)
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@180346 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/j3/VMCore/JavaConstantPool.cpp
@@ -508,10 +508,10 @@ JavaField* JavaConstantPool::lookupField(uint32 index, bool stat) {
       // don't throw if no field, the exception will be thrown just in time  
       if (field) {
         if (!stat) {
-          ctpRes[index] = (void*)field->ptrOffset;
+          ctpRes[index] = (void*)(uintptr_t)field->ptrOffset;
         } else if (lookup->isReady()) {
           void* S = field->classDef->getStaticInstance();
-          ctpRes[index] = (void*)((uint64)S + field->ptrOffset);
+          ctpRes[index] = (void*)(uintptr_t)((uint64)S + field->ptrOffset);
         }
       }
       return field;
